requirements.txt
# Core ML/Data Science
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==2.0.0
joblib==1.3.2

# MLOps Tools - MLflow 2.x for DagHub
mlflow>=2.0.0,<3.0.0
dagshub
dvc==3.27.0
dvc-s3==3.0.1

# API & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=1.10.0,<2.0.0
python-multipart==0.0.6
requests==2.31.0

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Utilities
python-dotenv==1.0.0
boto3==1.29.7
minio==7.2.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1
python-dateutil==2.8.2

# Visualization
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.18.0